+++ sqry-cli/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [8.0.2](https://github.com/verivus-oss/sqry/compare/v8.0.1...v8.0.2) - 2026-04-11
+
+### Other
+- update Cargo.lock dependencies
+## [8.0.1](https://github.com/verivus-oss/sqry/compare/v8.0.0...v8.0.1) - 2026-04-11
+
+### Other
+- update Cargo.lock dependencies
 ## [8.0.0](https://github.com/verivus-oss/sqry/compare/v7.2.0...v8.0.0) - 2026-04-10
 
 ### Added